Improved the Features Section
Really when I first arrived, I did not understand what this caching plug in did.  It just looked like more complexity, which I did not need.  So it is good to call out the high level concepts embedded here, so that people have an instant idea of what this excellent plug in is capable of.   

Sorry if I made any mistakes, I am still new at this, but generally this is a much better high level introduction to what this plug in does.  It is hard to  get the abstractions right by just reading a configuration file. 

I hope this helps.